move handler functions in message.ts
