+++ src/handler/index.ts
@@ -1,5 +1,5 @@
 import { handleCountingEditedMessage } from './handleCountingEditedMessage';
-import { handleCountingMessage } from './handleCountingMessage';
+import { handleCountingMessage } from './message/handleCountingMessage';
 import { handleDeathOfNira } from './handleDeathOfNira';
 import { handleNiraWave } from './handleNiraWave';
 import { handleNoUMessage } from './handleNoUMessage';
@@ -10,7 +10,16 @@ import { handleNiraMessageReactions } from './handleNiraMessageReactions';
 import { handleCheckNiraMojis } from './handleCheckNiraMojis';
 import { handlePatPatRole } from './handlePatPatRole';
 import { handleNiraMojisInChannels } from './handleNiraMojisInChannels';
-import { handleWorkInUniguriMessage } from './handleWorkInUniguriMessage';
+import { handleWorkInUniguriMessage } from './message/handleWorkInUniguriMessage';
+import { handleUwuChannelMessage } from './message/handleUwuChannelMessage';
+import { handleBotCheckMessage } from './message/handleBotCheckMessage';
+import { handleGreyMessage } from './message/handleGreyMessage';
+import { handleWelcomeMessage } from './message/handleWelcomeMessage';
+import { handleSubscriptionThankYouMessage } from './message/handleSubscriptionThankYouMessage';
+import { handlePatPatCommandMessage } from './message/handlePatPatCommandMessage';
+import { handleNonNitroEmoji } from './message/handleNonNitroEmoji';
+import { handleServerRulesMessage } from './message/handleServerRulesMessage';
+import { handleFishyCommandsMessage, handleOtherFishyMessage } from './message/handleFishyCommandsMessage';
 
 export {
   handleCountingEditedMessage,
@@ -26,4 +35,14 @@ export {
   handlePatPatRole,
   handleNiraMojisInChannels,
   handleWorkInUniguriMessage,
+  handleUwuChannelMessage,
+  handleBotCheckMessage,
+  handleGreyMessage,
+  handleWelcomeMessage,
+  handleSubscriptionThankYouMessage,
+  handlePatPatCommandMessage,
+  handleNonNitroEmoji,
+  handleServerRulesMessage,
+  handleFishyCommandsMessage,
+  handleOtherFishyMessage,
 };
