+++ src/handler/message/handleWorkInUniguriMessage.ts
@@ -1,6 +1,6 @@
 import { Message } from 'discord.js';
 
-import { themechannels } from '../config/config.json';
+import { themechannels } from '../../config/config.json';
 
 export const handleWorkInUniguriMessage = (message: Message) => {
   if (message.channel.id !== themechannels.uniguri) {
